+++ src/main/java/pl/schoolmanagementsystem/common/teacher/TeacherInClassRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface TeacherInClassRepository extends JpaRepository<TeacherInClass, Integer> {
+public interface TeacherInClassRepository extends JpaRepository<TeacherInClass, Long> {
 
     @Query("select t from TeacherInClass t left join fetch t.taughtClasses where t.teacher=?1 and t.taughtSubject=?2")
     Optional<TeacherInClass> findByTeacherAndTaughtSubject(Teacher teacher, String schoolSubject);
